requirements.txt
django
djangorestframework
environs
mysqlclient
gunicorn
psycopg2